+++ src/main/java/com/soo/routine/repository/MissionIconRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MissionIconRepository extends JpaRepository<MissionIcon, Integer> {
+public interface MissionIconRepository extends JpaRepository<MissionIcon, Long> {
     
 }
